[APDV-111] Moved GroupEndpointMapper
